requirements.txt
python-telegram-bot[ext]
psycopg2-binary
python-dotenv
google-generativeai
Pillow
httpx